fixes some typos
