+++ src/components/common/detail/detailHeader/index.js
@@ -13,10 +13,9 @@ const useStyles = makeStyles({
     textTransform: 'uppercase',
   },
   toolbar: {
-    display: 'flex',
-  },
-  tool: {
-    margin: '0 .5em',
+    display: 'grid',
+    gridTemplateColumns: '1fr 1fr',
+    gridColumnGap: 15,
   },
 }, { name: 'Mui_Styles_DetailHeader' });
 
@@ -28,16 +27,13 @@ const DetailHeader = props => {
     <header className={classes.root}>
       <div className={classes.label}>
         {title}
-        {isRequired ? ' *' : ''}
+        {isRequired ? ' *' : null}
       </div>
-      {children
-        && (
+      {children && (
         <div className={classes.toolbar}>
-          <div className={classes.tool}>
-            {children}
-          </div>
+          {children}
         </div>
-        )}
+      )}
     </header>
   );
 };
